+++ src/main/java/fr/inrae/act/bagap/chloe/window/kernel/entity/EntityCountValueAndCoupleKernel.java
@@ -10,7 +10,7 @@ public class EntityCountValueAndCoupleKernel extends EntityLandscapeMetricKernel
 	
 	private int[] lastEntityLine;
 	
-	private short[] lastValueLine;
+	private int[] lastValueLine;
 
 	public EntityCountValueAndCoupleKernel(int noDataValue, int[] values){
 		super(noDataValue);
@@ -45,20 +45,20 @@ public class EntityCountValueAndCoupleKernel extends EntityLandscapeMetricKernel
 	@Override
 	public void init(){
 		lastEntityLine = new int[width()];
-		lastValueLine = new short[width()];
+		lastValueLine = new int[width()];
 	}
 	
 	@Override
 	public void applyEntityWindow(){
 		
 		int mv, va;
-		short v=0, v_H, v_V;
+		int v=0, v_H, v_V;
 		for(int y=0; y<height(); y++){
 			for(int x=0; x<width(); x++){
 				va = (int) entityDatas()[y*width() + x];
 				if(va != 0 && va != noDataValue()){
 					
-					v = (short) inDatas()[y*width() + x];
+					v = (int) inDatas()[y*width() + x];
 					outDatas().get(va)[2] += 1;
 					
 					if(v == noDataValue()){
@@ -73,7 +73,7 @@ public class EntityCountValueAndCoupleKernel extends EntityLandscapeMetricKernel
 					// couple vertical
 					if((y > 0) && entityDatas()[(y-1)*width() + x] == va){
 					
-						v_V = (short) inDatas()[(y-1)*width() + x];
+						v_V = (int) inDatas()[(y-1)*width() + x];
 						outDatas().get(va)[nbValues+5] += 1;
 						
 						if(v == noDataValue() || v_V == noDataValue()){
@@ -102,7 +102,7 @@ public class EntityCountValueAndCoupleKernel extends EntityLandscapeMetricKernel
 					// couple horizontal
 					if((x > 0) && entityDatas()[y*width() + (x-1)] == va){
 					
-						v_H = (short) inDatas()[y*width() + (x-1)];
+						v_H = (int) inDatas()[y*width() + (x-1)];
 						outDatas().get(va)[nbValues+5] += 1;
 						
 						if(v == noDataValue() || v_H == noDataValue()){
