+++ src/main/java/fr/inrae/act/bagap/chloe/window/kernel/grid/GridCountCoupleKernel.java
@@ -44,21 +44,21 @@ public class GridCountCoupleKernel extends GridLandscapeMetricKernel {
 			outDatas()[x][i] = 0f;
 		}
 			
-		short v, v_H, v_V;
+		int v, v_H, v_V;
 		int mc;				
 		for(int y=0; y<gridSize(); y++) {
 			if((theY+y) < height()){
 				for(int lx=0; lx<gridSize(); lx++) {
 					if((x*gridSize() + lx) < width()){
 						
-						v = (short) inDatas()[((theY+y)*width()) + (x*gridSize() + lx)];
+						v = (int) inDatas()[((theY+y)*width()) + (x*gridSize() + lx)];
 						outDatas()[x][2] += 1;
 						if(v == noDataValue()){
 							outDatas()[x][3] += 1;
 						}
 						
 						if(y > 0) {
-							v_V = (short) inDatas()[((theY+y-1)*width()) + (x*gridSize() + lx)];
+							v_V = (int) inDatas()[((theY+y-1)*width()) + (x*gridSize() + lx)];
 							outDatas()[x][4] += 1;
 							if(v == noDataValue() || v_V == noDataValue()){
 								outDatas()[x][5] += 1;
@@ -71,7 +71,7 @@ public class GridCountCoupleKernel extends GridLandscapeMetricKernel {
 						}
 						
 						if(lx > 0) {
-							v_H = (short) inDatas()[((theY+y)*width()) + (x*gridSize() + lx - 1)];
+							v_H = (int) inDatas()[((theY+y)*width()) + (x*gridSize() + lx - 1)];
 							outDatas()[x][4] += 1;
 							if(v == noDataValue() || v_H == noDataValue()){
 								outDatas()[x][5] += 1;
