+++ src/main/java/fr/inrae/act/bagap/chloe/window/kernel/selected/SelectedCountCoupleKernel.java
@@ -50,7 +50,7 @@ public class SelectedCountCoupleKernel extends SelectedLandscapeMetricKernel {
 			
 		final int mid = windowSize() / 2;
 		int ic, ic_V, ic_H;
-		short v, v_H, v_V;
+		int v, v_H, v_V;
 		int mc;
 		float coeff;
 		for (int dy = -mid; dy <= mid; dy += 1) {
@@ -60,7 +60,7 @@ public class SelectedCountCoupleKernel extends SelectedLandscapeMetricKernel {
 						ic = ((dy+mid) * windowSize()) + (dx+mid);
 						coeff = coeff()[ic];
 						if(coeff > 0){
-							v = (short) inDatas()[((y + dy) * width()) + (x + dx)];
+							v = (int) inDatas()[((y + dy) * width()) + (x + dx)];
 							outDatas().get(p)[2] += coeff;
 							if(v == noDataValue()){
 								outDatas().get(p)[3] += coeff;
@@ -68,7 +68,7 @@ public class SelectedCountCoupleKernel extends SelectedLandscapeMetricKernel {
 							if((dy > -mid) && ((y + dy) > 0)) {
 								ic_V = ((dy+mid-1) * windowSize()) + (dx+mid);
 								if(coeff()[ic_V] > 0){
-									v_V = (short) inDatas()[((y + dy - 1) * width()) + (x + dx)];
+									v_V = (int) inDatas()[((y + dy - 1) * width()) + (x + dx)];
 									outDatas().get(p)[4] += coeff;
 									if(v == noDataValue() || v_V == noDataValue()){
 										outDatas().get(p)[5] += coeff;
@@ -84,7 +84,7 @@ public class SelectedCountCoupleKernel extends SelectedLandscapeMetricKernel {
 							if((dx > -mid) && ((x + dx) > 0)) {
 								ic_H = ((dy+mid) * windowSize()) + (dx+mid-1);
 								if(coeff()[ic_H] > 0){
-									v_H = (short) inDatas()[((y + dy) * width()) + (x + dx - 1)];
+									v_H = (int) inDatas()[((y + dy) * width()) + (x + dx - 1)];
 									outDatas().get(p)[4] += coeff;
 									if(v == noDataValue() || v_H == noDataValue()){
 										outDatas().get(p)[5] += coeff;
