+++ src/main/java/fr/inrae/act/bagap/chloe/window/kernel/selected/SelectedCountValueAndCoupleKernel.java
@@ -52,7 +52,7 @@ public class SelectedCountValueAndCoupleKernel extends SelectedLandscapeMetricKe
 				
 		final int mid = windowSize() / 2;
 		int ic, ic_V, ic_H;
-		short v, v_H, v_V;
+		int v, v_H, v_V;
 		int mv;
 		float coeff;
 		for (int dy = -mid; dy <= mid; dy += 1) {
@@ -62,7 +62,7 @@ public class SelectedCountValueAndCoupleKernel extends SelectedLandscapeMetricKe
 						ic = ((dy+mid) * windowSize()) + (dx+mid);
 						coeff = coeff()[ic];
 						if(coeff > 0){
-							v = (short) inDatas()[((y + dy) * width()) + (x + dx)];
+							v = (int) inDatas()[((y + dy) * width()) + (x + dx)];
 							outDatas().get(p)[2] += coeff;
 							
 							if(v == noDataValue()){
@@ -77,7 +77,7 @@ public class SelectedCountValueAndCoupleKernel extends SelectedLandscapeMetricKe
 							if((dy > -mid) && ((y + dy) > 0)) {
 								ic_V = ((dy+mid-1) * windowSize()) + (dx+mid);
 								if(coeff()[ic_V] > 0){
-									v_V = (short) inDatas()[((y + dy - 1) * width()) + (x + dx)];
+									v_V = (int) inDatas()[((y + dy - 1) * width()) + (x + dx)];
 									outDatas().get(p)[nbValues+5] += coeff;
 									
 									if(v == noDataValue() || v_V == noDataValue()){
@@ -96,7 +96,7 @@ public class SelectedCountValueAndCoupleKernel extends SelectedLandscapeMetricKe
 							if((dx > -mid) && ((x + dx) > 0)) {
 								ic_H = ((dy+mid) * windowSize()) + (dx+mid-1);
 								if(coeff()[ic_H] > 0){
-									v_H = (short) inDatas()[((y + dy) * width()) + (x + dx - 1)];
+									v_H = (int) inDatas()[((y + dy) * width()) + (x + dx - 1)];
 									outDatas().get(p)[nbValues+5] += coeff;
 									
 									if(v == noDataValue() || v_H == noDataValue()){
