+++ src/main/java/fr/inrae/act/bagap/chloe/window/metric/patch/MeanPatchSizeIndex.java
@@ -12,9 +12,9 @@ public class MeanPatchSizeIndex extends Metric implements PatchMetric {
 	@Override
 	protected void doCalculate(Counting co) {
 		if(co.validValues() > 0){
-			int nbPatch = co.nbPatches();
+			double nbPatch = co.nbPatches();
 			if(nbPatch > 0){
-				value = (double) co.totalSurface()/nbPatch;
+				value = co.totalSurface()/nbPatch;
 			}else{
 				value = 0;
 			}
