+++ src/main/java/fr/inrae/act/bagap/chloe/window/metric/patch/MeanPatchSizeMultipleIndex.java
@@ -23,13 +23,13 @@ public class MeanPatchSizeMultipleIndex extends Metric implements PatchMetric {
 	protected void doCalculate(Counting co) {
 		if(co.validValues() > 0){
 			double totalSurface = 0.0;
-			int nbPatches = 0;
+			double nbPatches = 0;
 			for(short vv : v){
 				totalSurface += co.totalSurface(vv);
 				nbPatches += co.nbPatches(vv);
 			}
 			if(nbPatches > 0){
-				value = (double) totalSurface / nbPatches;
+				value = totalSurface / nbPatches;
 			}else{
 				value = 0;
 			}
